requirements.txt
Flask==1.1.2
pyjwt==1.7.1
